cleanup: remove dead code, dead schema fields, barrel files, dedup shared logic
