--- src/hooks/delegate-task-retry/index.ts
@@ -1,7 +1,0 @@
-export { buildRetryGuidance } from './guidance';
-export { createDelegateTaskRetryHook } from './hook';
-export type { DelegateTaskErrorPattern, DetectedError } from './patterns';
-export {
-  DELEGATE_TASK_ERROR_PATTERNS,
-  detectDelegateTaskError,
-} from './patterns';
