--- src/hooks/json-error-recovery/index.ts
@@ -1,6 +1,0 @@
-export {
-  createJsonErrorRecoveryHook,
-  JSON_ERROR_PATTERNS,
-  JSON_ERROR_REMINDER,
-  JSON_ERROR_TOOL_EXCLUDE_LIST,
-} from './hook';
