+++ src/utils/index.ts
@@ -1,8 +1,7 @@
 export * from './agent-variant';
 export * from './background-job-board';
-export * from './env';
 export * from './internal-initiator';
-export { getLogDir, initLogger, log, resetLogger } from './logger';
+export { getLogDir, initLogger, log } from './logger';
 export * from './polling';
 export * from './session';
 export * from './task';
